+++ wasmsdk/blobber.go
@@ -269,7 +269,7 @@ func Share(allocationID, remotePath, clientID, encryptionPublicKey string, expir
 	}
 
 	var fileName string
-	_, fileName = common.Split(remotePath)
+	_, fileName = pathutil.Split(remotePath)
 
 	if revoke {
 		err := allocationObj.RevokeShare(remotePath, clientID)
@@ -478,7 +478,7 @@ func uploadWithJsFuncs(allocationID, remotePath string, readChunkFuncName string
 	}
 	remotePath = zboxutil.GetFullRemotePath(localPath, remotePath)
 
-	_, fileName := common.Split(remotePath)
+	_, fileName := pathutil.Split(remotePath)
 
 	fileMeta := sdk.FileMeta{
 		Path:       localPath,
@@ -557,7 +557,7 @@ func upload(allocationID, remotePath string, fileBytes, thumbnailBytes []byte, e
 	}
 	remotePath = zboxutil.GetFullRemotePath(localPath, remotePath)
 
-	_, fileName := common.Split(remotePath)
+	_, fileName := pathutil.Split(remotePath)
 
 	fileMeta := sdk.FileMeta{
 		Path:       localPath,
