+++ zboxcore/sdk/allocation_file_test.go
@@ -183,7 +183,7 @@ func TestAllocation_UpdateFile(t *testing.T) {
 				"allocation_id":    a.ID,
 				"lookup_hash":      "lookup_hash",
 				"name":             mockLocalPath,
-				"path":             common.Join("/", mockLocalPath),
+				"path":             pathutil.Join("/", mockLocalPath),
 				"path_hash":        "path_hash",
 				"parent_path":      "/",
 				"level":            1,
@@ -288,7 +288,7 @@ func TestAllocation_UpdateFileWithThumbnail(t *testing.T) {
 				"allocation_id":    a.ID,
 				"lookup_hash":      "lookup_hash",
 				"name":             mockLocalPath,
-				"path":             common.Join("/", mockLocalPath),
+				"path":             pathutil.Join("/", mockLocalPath),
 				"path_hash":        "path_hash",
 				"parent_path":      "/",
 				"level":            1,
@@ -396,7 +396,7 @@ func TestAllocation_EncryptAndUpdateFile(t *testing.T) {
 				"allocation_id":    a.ID,
 				"lookup_hash":      "lookup_hash",
 				"name":             mockLocalPath,
-				"path":             common.Join("/", mockLocalPath),
+				"path":             pathutil.Join("/", mockLocalPath),
 				"path_hash":        "path_hash",
 				"parent_path":      "/",
 				"level":            1,
@@ -508,7 +508,7 @@ func TestAllocation_EncryptAndUpdateFileWithThumbnail(t *testing.T) {
 				"allocation_id":    a.ID,
 				"lookup_hash":      "lookup_hash",
 				"name":             mockLocalPath,
-				"path":             common.Join("/", mockLocalPath),
+				"path":             pathutil.Join("/", mockLocalPath),
 				"path_hash":        "path_hash",
 				"parent_path":      "/",
 				"level":            1,
